split off DepositSafetyThreshold and tests
into separate file
