+++ channel/state/outcome/single-asset-exit.go
@@ -13,27 +13,6 @@ type SingleAssetExit struct {
 	Allocations Allocations
 }
 
-// DepositSafetyThreshold returns the amount of this asset that a user with
-// the specified interests must see on-chain before the safe recoverability of
-// their own deposts is guaranteed
-func (s SingleAssetExit) DepositSafetyThreshold(interests []types.Destination) *big.Int {
-	sum := big.NewInt(0)
-
-	for _, allocation := range s.Allocations {
-
-		for _, interest := range interests {
-			if allocation.Destination == interest {
-				// we have 'hit' a destination whose balances we are interested in protecting
-				return sum
-			}
-		}
-
-		sum.Add(sum, allocation.Amount)
-	}
-
-	return sum
-}
-
 // TotalAllocated returns the toal amount allocated, summed across all destinations (regardless of AllocationType)
 func (sae SingleAssetExit) TotalAllocated() *big.Int {
 	return sae.Allocations.Total()
